challenges/image-service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
Pillow>=10.3,<12.0
python-multipart>=0.0.9,<1.0

